+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## 1.3.1 - 2018-10-22
 ### Fixed
 - Fix environment variable interpolation when `WithDefault` is used.
 - Fix support for non-scalar keys in YAML mappings (eg: "1: car").